Refactor serialization and deserialization in schemas
